+++ logisland-components/logisland-processors/logisland-processor-common/src/main/java/com/hurence/logisland/processor/AddFields.java
@@ -15,31 +15,101 @@
  */
 package com.hurence.logisland.processor;
 
+import com.hurence.logisland.annotation.behavior.DynamicProperties;
 import com.hurence.logisland.annotation.behavior.DynamicProperty;
 import com.hurence.logisland.annotation.documentation.CapabilityDescription;
 import com.hurence.logisland.annotation.documentation.Tags;
 import com.hurence.logisland.component.AllowableValue;
 import com.hurence.logisland.component.PropertyDescriptor;
+import com.hurence.logisland.component.PropertyValue;
 import com.hurence.logisland.record.Field;
+import com.hurence.logisland.record.FieldType;
 import com.hurence.logisland.record.Record;
 import com.hurence.logisland.validator.StandardValidators;
+import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.util.*;
 
 @Tags({"record", "fields", "Add"})
-@CapabilityDescription("Add one or more field with a default value\n" +
-        "...")
-@DynamicProperty(name = "field to add",
-        supportsExpressionLanguage = false,
-        value = "a default value",
-        description = "Add a field to the record with the default value")
+@CapabilityDescription("Add one or more field with constant value or dynamic value using the `expression-language <./expression-language.html>`_." +
+        "Some examples of settings: \n\n" +
+        ".. code::\n" +
+        "\n" +
+        "\tnewStringField: bonjour\n" +
+        "\tnewIntField: 14\n" +
+        "\tnewIntField"+ AddFields.DYNAMIC_PROPS_TYPE_SUFFIX + ": INT\n" +
+        "\n" +
+        "Would add those fields in record :\n" +
+        "\n" +
+        ".. code::\n" +
+        "\n" +
+        "\tField{name='newStringField', type='STRING', value='bonjour'}\n" +
+        "\tField{name='newIntField', type='INT', value=14}\n" +
+        "\n" +
+        "Here a second example using expression language, once for the value, once for the key. Note that you can use for both." +
+        "We suppose that our record got already those fields : \n" +
+        "\n" +
+        ".. code::\n" +
+        "\n" +
+        "\tField{name='field1', type='STRING', value='bonjour'}\n" +
+        "\tField{name='field2', type='INT', value=14}\n" +
+        "\n" +
+        "This settings :\n" +
+        ".. code::\n" +
+        "\n" +
+        "\tnewStringField: ${field1 + \"-\" + field2}\n" +
+        "\tfieldToCalulateKey: 555\n" +
+        "\tfieldToCalulateKey"+ AddFields.DYNAMIC_PROPS_NAME_SUFFIX + ": ${\"_\" + field1 + \"-\"}\n" +
+        "\n" +
+        "Would add those fields in record :\n" +
+        "\n" +
+        ".. code::\n" +
+        "\n" +
+        "\tField{name='newStringField', type='STRING', value='bonjour-14'}\n" +
+        "\tField{name='_bonjour-', type='STRING', value='555'}\n" +
+        "\n\n" +
+        "As you probably notice, you can not add fields with name ending by either '" + AddFields.DYNAMIC_PROPS_NAME_SUFFIX + "' either '" +
+        AddFields.DYNAMIC_PROPS_TYPE_SUFFIX + "' because they are suffix are used to sort dynamic properties. But if you really want to do this a workaround" +
+        " is to specify the name of the field oui expression language, for example this settings would work:\n\n" +
+        ".. code::\n" +
+        "\n" +
+        "\tfieldWithReservedSuffix: bonjour\n" +
+        "\tfieldWithReservedSuffix"+ AddFields.DYNAMIC_PROPS_TYPE_SUFFIX + ": INT\n" +
+        "\tfieldWithReservedSuffix"+ AddFields.DYNAMIC_PROPS_TYPE_SUFFIX + ": myfield.endind.with.reserved.suffix"+AddFields.DYNAMIC_PROPS_TYPE_SUFFIX+"\n" +
+        "\n")
+
+@DynamicProperties(value = {
+        @DynamicProperty(name = "Name of the field to add",
+                supportsExpressionLanguage = true,
+                value = "Value of the field to add",
+                description = "Add a field to the record with the specified value. Expression language can be used." +
+                        "You can not add a field that end with '.type' as this suffix is used to specify the type of fields to add",
+                nameForDoc = "fakeField"),
+        @DynamicProperty(name = "Name of the field to add with the suffix '"+ AddFields.DYNAMIC_PROPS_TYPE_SUFFIX +"'",
+                supportsExpressionLanguage = false,
+                value = "Type of the field to add",
+                description = "Add a field to the record with the specified type. These properties are only used if a correspondant property without" +
+                        " the suffix '"+ AddFields.DYNAMIC_PROPS_TYPE_SUFFIX +"' is already defined. If this property is not defined, default type for adding fields is String." +
+                        "You can only use Logisland predefined type fields.",
+                nameForDoc = "fakeField" + AddFields.DYNAMIC_PROPS_TYPE_SUFFIX),
+        @DynamicProperty(name = "Name of the field to add with the suffix '" + AddFields.DYNAMIC_PROPS_NAME_SUFFIX + "'",
+                supportsExpressionLanguage = true,
+                value = "Name of the field to add using expression language",
+                description = "Add a field to the record with the specified name (which is evaluated using expression language). " +
+                        "These properties are only used if a correspondant property without" +
+                        " the suffix '" + AddFields.DYNAMIC_PROPS_NAME_SUFFIX + "' is already defined. If this property is not defined, " +
+                        "the name of the field to add is the key of the first dynamic property (which is the main and only required dynamic property).",
+                nameForDoc = "fakeField" + AddFields.DYNAMIC_PROPS_NAME_SUFFIX)
+        })
 public class AddFields extends AbstractProcessor {
 
 
     private static final Logger logger = LoggerFactory.getLogger(AddFields.class);
 
+    public static final String DYNAMIC_PROPS_TYPE_SUFFIX = ".field.type";
+    public static final String DYNAMIC_PROPS_NAME_SUFFIX = ".field.name";
 
     private static final AllowableValue OVERWRITE_EXISTING =
             new AllowableValue("overwrite_existing", "overwrite existing field", "if field already exist");
@@ -61,70 +131,108 @@ public class AddFields extends AbstractProcessor {
         return Collections.singletonList(CONFLICT_RESOLUTION_POLICY);
     }
 
+    private final Set<PropertyDescriptor> dynamicFieldProperties = new HashSet<>();
+    private final  Map<String, PropertyDescriptor> dynamicTypeProperties = new HashMap<>();
+    private final Map<String, PropertyDescriptor> dynamicNameProperties = new HashMap<>();
+    private String conflictPolicy;
 
     @Override
     protected PropertyDescriptor getSupportedDynamicPropertyDescriptor(final String propertyDescriptorName) {
+        if (propertyDescriptorName.endsWith(DYNAMIC_PROPS_TYPE_SUFFIX)) {
+            return new PropertyDescriptor.Builder()
+                    .name(propertyDescriptorName)
+                    .expressionLanguageSupported(false)
+                    .addValidator(new StandardValidators.EnumValidator(FieldType.class))
+                    .allowableValues(FieldType.values())
+                    .defaultValue(FieldType.STRING.getName().toUpperCase())
+                    .required(false)
+                    .dynamic(true)
+                    .build();
+        }
+        if (propertyDescriptorName.endsWith(DYNAMIC_PROPS_NAME_SUFFIX)) {
+            return new PropertyDescriptor.Builder()
+                    .name(propertyDescriptorName)
+                    .expressionLanguageSupported(true)
+                    .addValidator(StandardValidators.NON_EMPTY_VALIDATOR)
+                    .required(false)
+                    .dynamic(true)
+                    .build();
+        }
         return new PropertyDescriptor.Builder()
                 .name(propertyDescriptorName)
                 .expressionLanguageSupported(true)
-                .addValidator(StandardValidators.COMMA_SEPARATED_LIST_VALIDATOR)
+                .addValidator(StandardValidators.NON_EMPTY_VALIDATOR)
                 .required(false)
                 .dynamic(true)
                 .build();
     }
 
+
+    @Override
+    public void init(ProcessContext context) {
+        super.init(context);
+        initDynamicProperties(context);
+        this.conflictPolicy = context.getPropertyValue(CONFLICT_RESOLUTION_POLICY).asString();
+    }
+
     @Override
     public Collection<Record> process(ProcessContext context, Collection<Record> records) {
-        Map<String, String> fieldsNameMapping = getFieldsNameMapping(context);
         for (Record record : records) {
-            updateRecord(context, record, fieldsNameMapping);
+            updateRecord(context, record);
         }
         return records;
     }
 
-
-    private void updateRecord(ProcessContext context, Record record, Map<String, String> fieldsNameMapping) {
-
-        String conflictPolicy = context.getPropertyValue(CONFLICT_RESOLUTION_POLICY).asString();
-
-        if ((fieldsNameMapping == null) || (fieldsNameMapping.keySet() == null)) {
-            return;
-        }
-        fieldsNameMapping.keySet().forEach(addedFieldName -> {
-            final String defaultValueToAdd = context.getPropertyValue(addedFieldName).evaluate(record).asString();
-            // field is already here
-            if (record.hasField(addedFieldName)) {
-                if (conflictPolicy.equals(OVERWRITE_EXISTING.getValue())) {
-                    overwriteObsoleteFieldValue(record, addedFieldName, defaultValueToAdd);
-                }
-            } else {
-                record.setStringField(addedFieldName, defaultValueToAdd);
+    private void updateRecord(ProcessContext context, Record record) {
+        dynamicFieldProperties.forEach(addedFieldDescriptor -> {
+            final PropertyValue propValueOfValue = context.getPropertyValue(addedFieldDescriptor).evaluate(record);
+            String fieldNameToAdd = addedFieldDescriptor.getName();
+            if (dynamicNameProperties.containsKey(fieldNameToAdd)) {
+                fieldNameToAdd = context.getPropertyValue(dynamicNameProperties.get(fieldNameToAdd)).evaluate(record).asString();
+            }
+            if (!record.hasField(fieldNameToAdd) || conflictPolicy.equals(OVERWRITE_EXISTING.getValue())) {
+                setFieldValue(record, fieldNameToAdd, propValueOfValue, context);
             }
         });
     }
 
-    private void overwriteObsoleteFieldValue(Record record, String fieldName, String newValue) {
-        final Field fieldToUpdate = record.getField(fieldName);
-        record.removeField(fieldName);
-        record.setField(fieldName, fieldToUpdate.getType(), newValue);
+    private void setFieldValue(Record record, String fieldName, PropertyValue newValue, ProcessContext context) {
+        if (dynamicTypeProperties.containsKey(fieldName) && dynamicTypeProperties.get(fieldName).isDynamic()) {// un type été configuré
+            FieldType fieldType = FieldType.valueOf(context.getPropertyValue(dynamicTypeProperties.get(fieldName)).asString());
+            record.setField(fieldName, fieldType, newValue.getRawValue());
+        } else {// pas de type configuré
+            if (record.hasField(fieldName)) { //utilise l'ancien type
+                updateRecord(record, fieldName, newValue);
+            } else {//utilise le type String par défaut
+                record.setStringField(fieldName, newValue.asString());
+            }
+        }
     }
 
-    private Map<String, String> getFieldsNameMapping(ProcessContext context) {
-        /**
-         * list alternative regex
-         */
-        Map<String, String> fieldsNameMappings = new HashMap<>();
+    private void updateRecord(Record record, String fieldName, PropertyValue newValue) {
+        final Field fieldToUpdate = record.getField(fieldName);
+        record.removeField(fieldName);
+        record.setField(fieldName, fieldToUpdate.getType(), newValue.getRawValue());
+    }
+
+    private void initDynamicProperties(ProcessContext context) {
+        dynamicFieldProperties.clear();
+        dynamicTypeProperties.clear();
+        dynamicNameProperties.clear();
         // loop over dynamic properties to add alternative regex
         for (final Map.Entry<PropertyDescriptor, String> entry : context.getProperties().entrySet()) {
             if (!entry.getKey().isDynamic()) {
                 continue;
             }
-
-            final String fieldName = entry.getKey().getName();
-            final String mapping = entry.getValue();
-
-            fieldsNameMappings.put(fieldName, mapping);
+            if (entry.getKey().getName().endsWith(DYNAMIC_PROPS_TYPE_SUFFIX)) {
+                dynamicTypeProperties.put(StringUtils.removeEnd(entry.getKey().getName(), DYNAMIC_PROPS_TYPE_SUFFIX), entry.getKey());
+                continue;
+            }
+            if (entry.getKey().getName().endsWith(DYNAMIC_PROPS_NAME_SUFFIX)) {
+                dynamicNameProperties.put(StringUtils.removeEnd(entry.getKey().getName(), DYNAMIC_PROPS_NAME_SUFFIX), entry.getKey());
+                continue;
+            }
+            dynamicFieldProperties.add(entry.getKey());
         }
-        return fieldsNameMappings;
     }
 }