+++ logisland-components/logisland-processors/logisland-processor-common/src/main/java/com/hurence/logisland/processor/alerting/CheckThresholds.java
@@ -29,7 +29,7 @@ import javax.script.ScriptException;
 import java.util.*;
 
 @Tags({"record", "threshold", "tag", "alerting"})
-@CapabilityDescription("Compute threshold cross from given formulas.\n" +
+@CapabilityDescription("Compute threshold cross from given formulas.\n\n" +
         "- each dynamic property will return a new record according to the formula definition\n" +
         "- the record name will be set to the property name\n" +
         "- the record time will be set to the current timestamp\n")