+++ logisland-components/logisland-processors/logisland-processor-common/src/test/java/com/hurence/logisland/processor/ExpandMapFieldsTest.java
@@ -54,7 +54,7 @@ public class ExpandMapFieldsTest {
         runner.enqueue(createTestRecord());
         runner.run();
         runner.assertOutputRecordsCount(1);
-        runner.assertAllRecords(record -> {
+        runner.assertAllOutputRecords(record -> {
             Assert.assertFalse(record.hasField("map"));
             Assert.assertTrue(record.hasField("f1"));
             Assert.assertTrue(record.hasField("f2"));
@@ -80,7 +80,7 @@ public class ExpandMapFieldsTest {
         runner.enqueue(createTestRecord());
         runner.run();
         runner.assertOutputRecordsCount(1);
-        runner.assertAllRecords(record -> {
+        runner.assertAllOutputRecords(record -> {
             Assert.assertFalse(record.hasField("map"));
             Assert.assertTrue(record.hasField("f1"));
             Assert.assertTrue(record.hasField("f2"));